Make applyConfig transform argv instead of builders (#84)
Per the `yargs` docs, [`yargs.middleware(...)`](https://yargs.js.org/docs/#api-reference-middlewarecallbacks-applybeforevalidation) is intended to be used for `argv` transformations that should be run before a command is applied.

`applyConfig` used to transform the object exported by `builders`. This is a bad idea. Instead, it now transforms `argv` in the `yargs` middleware, just like `sanitizeInputs` was already doing.

Additionally, `assignGlobals` was renamed to `setSnapGlobals`.