Move `isRunning` check from startPlugin to _startPlugin (#21)
`_startPlugin` is the private wrapper function used to start plugins in the `PluginController`. This PR moves the check for whether the requested plugin is already running from the public `startPlugin` to `_startPlugin`, so that this check is always done. A test case is added to test this check.

While we're here, also fixes plugin controller state initialization and makes the state constructor argument property optional.