Lower minimum Node version to 12.11.0 (#125)
This PR lowers the minimum Node version to 12.11.0 in the manifest `engines` field, and sets the version in the `.nvmrc` to `v12`.

We increased the minimum Node version to 14 in #77, unfortunately for unknown reasons (`:spiderman-pointing-gif:`). Ideally, we keep it working with all current LTS versions. Since `worker_threads` [first became stable](https://nodejs.org/en/blog/release/v12.11.0/) in 12.11.0, we can't go below that, but that's fine.