Update Snap initial states (#126)
This PR changes the "initial" snap state (per the snap status state machine config) from `idle` to `installing`, and changes the state of rehydrated snaps to `stopped`.

When the Snap controller is initialized on application startup, no snap will be running. This state of "not running" is no different from that of a snap that has been shut down after not processing any requests for too long, and it is therefore appropriate to use `stopped` in both case.

This leaves the only case for the "initial" state as the period between the serialized snap object being added to controller state and it being started for the first time after its initial permissions have been approved. The snap is not "idle" during this time, but the application is making the necessary changes to execute it. Therefore, `installing` is a better name for this state. As a bonus, we can now unambiguously refer to snaps that are about to be shut down due to not processing any requests as "idle", even though this is not an actual status (they go straight from `running` to `stopped` in such cases, and we say that "idle" snaps should be stopped).