Improve types and documentation for `selectHooks`
The `selectHooks` function now has improved types and documentation.
Previously it had no JSDoc comments at all. Now the inline docs written
for the extension `selectHooks` function by @rekmarks has been ported
here, and docs were added for the generic parameters as well.

The types were a bit of a mess previously. There were multiple type
signatures, for one thing. The generic parameters seemed backwards,
because the keys for `hooks` were derived from the names to select,
even though the names to select are a _subset_ of the hook keys. The
generic parameters were poorly named. And there was some invalid type
casting.

Unfortunately we still need _three_ type casts here because of how
unfriendly `.reduce` is to work with in TypeScript, but at least all
of these casts are guaranteed to be correct.