Add "Endowment" permissions (#152)
Introduces the notion of permission types to the `PermissionController`, and adds "Endowment" permissions. Invoking these permissions required a new method and action, `PermissionController.getEndowments` and `PermissionController:getEndowments`, respectively. See the updated `PermissionController` README and linked issue for details.

Here is the description of endowment permissions, from said README:

The name "endowment" comes from the endowments that you may provide to a [Secure EcmaScript (SES) `Compartment`](https://github.com/endojs/endo/tree/26d991afb01cf824827db0c958c50970e038112f/packages/ses#compartment) when it is constructed.
SES endowments are simply names that appear in the compartment's global scope.
In the context of the `PermissionController`, endowments are simply "things" that subjects should not be able to access by default.
They _could_ be the names of endowments that are to be made available to a particular SES `Compartment`, but they could also be any JavaScript value, and it is the host's responsibility to make sense of them.

At present, endowment permissions may not have any caveats, but caveat support may be added in the future.

Closes #137