Add network access endowment (#154)
Adds a network access endowment permission. Currently it just sits in the same directory as the `PermissionController`, but we will move into a new package (`@metamask/permission-utils` or something to that effect) in the future.

This endowment is extremely simple; it takes no arguments, and its getter just returns `['fetch']`.