Use endowment permissions in SnapController (#155)
This PR adds support endowment permissions to the `SnapController`. We expect all endowments to be arrays of strings, and will have to update the endowments integration if that assumption is violated. It should be fine for now though. Endowment support as already added to the iframe execution environment here: https://github.com/MetaMask/iframe-execution-environment/pull/39

Support for the `endowments` property in `SnapExecutionData` will be added to the `SnapWorker` in this repository in a follow-up.