Update `snap_confirm` parameters (#158)
During review of the implementation, we discovered some issues with the naming of the parameters for `snap_confirm` method. This PR converts the parameters to an object format, updates their names and descriptions, and adds validation for each field. Each field remains a string, and the maximum length of each field was chosen after experimenting with various text lengths in the extension UI. The `textAreaContent` length of 1800 was chosen because this is roughly the character count of a single, U.S. letter-sized page of double-spaced 12pt Times New Roman, i.e. way more than anyone is going to read in practice anyway.

Co-authored-by: Shane <jonas.shane@gmail.com>