Add Snap installation events (#162)
This PR adds installation-related `ControllerMessenger` events to the `SnapController`, fixes a bug in `installSnaps`, makes some methods private, and deletes `runExistingSnaps`.

The following events are added:
- `snapAdded`, emitted when a Snap has been added to state and its status is `installing`
  - We will leverage this event to record metadata about the Snap in the extension.
- `snapInstalled`, emitted when a Snap has been fully installed
- `snapRemoved`, emitted when a Snap is removed

A bug was introduced in #140, where a successfully installed Snap would be overwritten with an unrelated error in the `installSnaps` result object. This PR prevents this overwrite from happening.

The `authorize` and `processRequestedSnap` methods are made private, since they should never be called by third parties. In truth, `add` should also be private, but we currently rely on it for the majority of our tests, and will attend to it later.

Finally, the `runExistingSnaps` method is removed because it's a hangover from the old Snaps lifecycle.