Add Snap SVG icon support (#163)
Adds support for the Snap [manifest `iconPath` field](https://github.com/MetaMask/specifications/blob/d4a5bf5d6990bb5b02a98bd3f95a24ffb28c701c/snaps/publishing.md#iconpath), and storing SVG icons as part of Snap state.

The `iconPath` field is the string path to an `.svg` file from the project root directory. If we detect this field, we read the corresponding field and store it as a string in the Snap object. This allows us to display the Snap's icon in the MetaMask UI. If reading the file fails, we throw a validation error.