Fix build and publish scripts (#169)
While publishing `0.6.0`, I discovered a number of issues with the package-wide build and publishing scripts. This PR fixes these issues by adding a `build:pre-tsc` and `build:post-tsc` script to each package, and ensuring that these get executed for each package in the correct order at the time of publishing.