Fix `mm-snap init` `src` default value (#186)
The `src` value (i.e. `packageJson.main`) should default to `src/index.js`, however we were defaulting to `index.js` because that is the default in the package we use to build the `package.json` file, `init-package-json`. This PR ensures that this value defaults to `src/index.js` by means of a one-line patch to `init-package-json`. `patch-package` was added as a `devDependency` to support this.