Fix `mm-snap build` command when CLI is installed globally (#216)
As reported in #206, `mm-snap build` fails with an error reporting that `babelify` cannot be found if the CLI is installed globally. In local testing, I can reproduce this behavior with both `yarn` and `npm`. The cause of this error has something to do with the way Browserify and / or Babel resolve dynamically imports their various transforms, presets, and plugins. By importing all of these statically using `require` instead of passing their names as string literals, the "module not found" error disappears.

I considered bundling the CLI as an alternative, but that appears to be difficult precisely due to the dynamic imports that caused the problem in the first place.