Add snaps-cli usage instructions to its readme (#228)
Adds the output of `mm-snap --help` to the `@metamask/snaps-cli` readme. We used to keep it in the readme but I removed it because we were forgetting to update it manually. Then I realized that I am able to program computers to automate menial tasks for me, so now we just do that instead!