Add optional HTML comment transform (#237)
A user recently reported that their snap bundle was rejected by SES due to HTML-like comments appearing in a transitive dependency. This PR restores our aggressive handling of this problem by default, with an option of disabling it. We will follow up this PR with the ability to add user-specified transforms to the `mm-snap build` process.