Default to verboseErrors, fix global options (#249)
This PR changes the default value of `verbosErrors` to `true`, under the assumption that we should surface as much information to the user as possible by default.

In making this change, I discovered a bug with how we assign our global options (e.g. `verboseErrors`), which is also fixed by this PR. Specifically, we were calling the `Boolean` constructor on the value of each global option to convert them to actual booleans on the `globalThis.snaps` object, but that would of course set any non-empty string to `true`. It's possible that this "bug" only affects our tests due to how `yargs` parses input, but handling this ourselves is easy and cheap.