Specify all endowments in the `SnapController` (#266)
In #252, started to specify the list of snap endowments in the `SnapController` instead of the `BaseSnapExecutor`. Notably exceptions to this included any endowment that we needed to modify in the `BaseSnapExecutor` as opposed to pull straight from the global object, such as `setTimeout` and `clearTimeout`. This PR completes the effort we started in #252 by adding logic for creating modified / attenuated endowments _in response_ to them being in the list of endowments received by the executor (i.e., the list we specify in the `SnapController` and constructor from our defaults and the snap's permissions).

The only endowment that's still specified by default in the executor is our `wallet` object, which, well, we gotta have that.

The individual changes to each affected package (`snap-controllers` and `execution-environments`) are breaking, but there will be no behavior our breakage in the extension or iframe execution environment when we update these packages all at once.