Fix comment stripping bug (#270)
While working with a snap locally I discovered that `@nodefactory/strip-comments` fails to handle block comments of the form `/**/`. When it encounters this string, it will remove the `/**/` token and the entire remainder of the string being operated on. This PR fixes this bugs by manually stripping block comments of this form.