Populate `jsonrpc` field in Snap RPC requests (#273)
This PR ensures that the `jsonrpc` field of JSON-RPC request objects is valid before forwarding the object to the Snap via its RPC handler. If the field is missing entirely, we simply populate it with the string `2.0`. If the field exists and has any other value than the string `2.0`, we throw an error.

This is change is due to more stringent validation of these requests added to execution environments in #231. This is probably a good thing, so we update the `SnapController` to match.