Stop binding function endowments to root globalThis (#294)
This PR fixes function endowments with properties in our Snap execution environments. Previously, values with a type of `"function"` would be added to the endowments object as `value.bind(globalThis)`. We failed to realize that this would erase the own properties of values like `Date`, leading to e.g. `Date.now` being undefined inside the child Compartment. We really should `harden` the entire endowments object, but this causes tests to fail. In order to resolve the immediate problem, we simply add all global values to the endowments object as-is for now.