Remove staged linting for changelogs (#455)
The staged linting for changelogs fails on my device when multiple changelog files are staged. I suspect the `--file` flag of the `auto-changelog` CLI does not accept multiple files.

Even if that worked, we would have to conditionally pass the lint command the `--rc` flag for releases in order for it to pass / be useful, so the simplest solution seemed to be to remove it.