Close timer handle leak
Platform timer handles leak authority. On the web, they leak the authority to cancel any timer, in this case any registered timer. On Node.js, they leak the authority to manipulate the timer schedule data structures arbitrarily. With this change, guest code receives an opaque token that can be safely passed as the authority to cancel any single timer.

An analogous change would close a leak in interval timers.