Strip leading `./` from paths in npm snap manifests (#537)
Strips the leading `./` from the values of path fields in `snap.manifest.json` of npm snaps. Although this is not [according to spec](https://github.com/MetaMask/specifications/blob/main/snaps/publishing.md#filepath), we do it by the principle of "be lenient about what you receive and stringent about what you send out". We do no extend this courtesy for locally developed snaps because it's better for the issues to be caught and addressed.