Rename `onRpcMessage` to `onRpcRequest` (#538)
Renames the Snap `onRpcMessage` export to `onRpcRequest`. The underlying type, `SnapRpcHandler` takes a single named parameter object with two properties, `origin` and `request`. Since the "message" is a "request" both by the parameter name and the JSON-RPC 2.0 specification (even if the request is technically a notification), we should use "request" for the export name as well.