Rename package `publish` command to `publish:package` (#547)
Since the Yarn v3 migration, `yarn` and/or `npm` has run the `publish` script twice whenever `publish:all` is run from the root. The first time it succeeds, the second time it fails due to the `OTP` positional variable not being specified, leading me to believe that this is occurring due to some lifecycle / naming collision issue.

This PR renames the `publish` script in workspace `package.json` files to `publish:package` in order to stop `yarn` from running the script twice. I haven't tested this, but it (1) can't hurt, (2) is less confusing because there's no naming collision with the `yarn` / `npm` CLI `publish` command, and (3) follows the convention established by the monorepo root `publish:all` script.