Add 'Access-Control-Allow-Origin': * to serve
When we try to install a snap on localhost metamask tries to fetch the snap.manifest.json file. Since the request comes from Metamask to a server that doesn’t respond with "Acces-Control-Allow-Origin" header the request is not allowed by the CORS browser policy.

I think that it should be fine to set 'Access-Control-Allow-Origin' to '*', becasue the cli and the serve command are intended for developer use and it shouldn’t be seen as a security issue.

fixes https://github.com/MetaMask/snaps-skunkworks/issues/635