Add WebWorkerExecutionService, WebWorkerPostMessageStream to snaps-controllers package
This commit adds a new service that will allow for code execution in a
Web Worker, along with the necessary stream to communicate between the
main thread and the worker. The service takes in a URL where the worker
pool document is hosted, and will lazily create and close the pool as
needed. The service also implements all of the necessary abstract
methods for the `AbstractExecutionService`. The WebWorkerPostMessageStream
wraps messages in a job ID, and forwards them to an underlying message
stream.