Add worker pool to the WebWorkerPool class
A pool of workers has been added to the WebWorkerPool class that allows for reusability and handling of many jobs without overloading the system. The pool creates new workers if the current pool size is below the minimum size. A worker can be retrieved from the pool using the #getWorker() method.