Update dependencies for snaps-controllers package & fix type errors
The "snap.manifest" files for several examples are updated to contain new shasum values. A function in serveUtils.ts is refactored to use the IncomingMessage type for the request parameter. The yarn.lock file is updated to reflect the same change. The type for the snaps property in NodeJS.Global is changed to be SnapsCliGlobals. Lastly, the version of "@types/readable-stream" is updated in the snaps-constrollers package.json file.