Synchronize package versions across the workspace (#1326)
* Add constraints on package dependency ranges

This commit adds utility predicates and constraints on the ranges specified in package dependencies. The changes ensure that all dependency ranges specified follow valid patterns, and that dependency ranges across the monorepo are synchronized. A dependency should not appear in both "dependencies" and "devDependencies".

* Synchronise package versions across all workspace packages

This fixes the constraints errors caused by the previous commit, by bringing all packages up to date to the latest used version.

* Add @babel/plugin-transform-modules-commonjs and @babel/helper-annotate-as-pure packages to policy.json

This commit adds the two mentioned packages to the policy.json file, allowing them to be used in the code.

* Update dependencies for snaps-controllers package & fix type errors

The "snap.manifest" files for several examples are updated to contain new shasum values. A function in serveUtils.ts is refactored to use the IncomingMessage type for the request parameter. The yarn.lock file is updated to reflect the same change. The type for the snaps property in NodeJS.Global is changed to be SnapsCliGlobals. Lastly, the version of "@types/readable-stream" is updated in the snaps-constrollers package.json file.

* Update snaps-cli dependencies and make CLI asynchronous

This commit updates the snaps-cli dependencies and changes the cli function to be asynchronous.

* Add dependency on "@metamask/providers"

This change adds the dependency "@metamask/providers" at version ^10.2.1 to the package. This was removed on accident after a rebase.

* Increase jest test timeout from 5 to 10 seconds

Some tests require more than 5 seconds to complete, increasing the timeout to 10 seconds should be enough to cover all test cases.

* Add @metamask/providers dependency to snaps-types

This change adds @metamask/providers package as a dependency to the project. This package is necessary to interact with provider engines.

* Add dependencies to fix peer dependencies warnings

This adds dependencies to different packages to resolve all peer dependencies warnings upon running Yarn.

* Revert "Update snaps-cli dependencies and make CLI asynchronous"

This reverts commit 573834bfacc11a94789d9add55a2a1c418ff017a.

* Update coverage branch, fix type issue

The branches threshold for the test coverage has been recalibrated. In the file `serveUtils.ts`, the type of IncomingMessage caused an error, so it has been fixed to match the required type.

* Disable eslint on unused expression and no-floating-promises

We disabled eslint on unused expression and no-floating-promises due to the use of yargs library in cli.ts.