Refactor build.js script
This change simplifies the code in the build.js script by consolidating the additional LavaMoat security options  into a single object. In addition, the conditional check for the `lavamoatSecurityOptions` has been modified to include cases where the html file is being built. Therefore, the additional options will also be included in html builds.