Prepare CI for Flask releases (#1327)
* Add release/1.0.0 branch to push event trigger

This change allows the workflow to trigger when changes are pushed to the main branch or release/1.0.0 branch.

* Add a script for preparing the package for a Flask build

This commit adds a script that checks if the current branch name is `main`. If it is, then
it executes a yarn script to prepare the package for a Flask build by overwriting the package name
to append a `-flask` suffix to the organization name.

* Add missing `then` to if statement

* Revert "Add missing `then` to if statement"

This reverts commit e2d65a899519ffa3f53da5453b2c84b90f37c46d.

* Revert "Add a script for preparing the package for a Flask build"

This reverts commit 6f69f6e4b06ff481827d0225632d1e33f88bf830.

* Add logic to handle releases for different branch

Add a new step to check if the push is on the main branch or not, if it is a push to the main branch, then the release tag will be 'flask', otherwise the release tag will be 'latest'. The tag is used when publishing the package.

* Check the current branch to determine which release tag to generate

By getting the name of the current branch and checking if it is `main`, this commit determines whether to create a release tag called "flask" or "latest". The change also explicitly sets the shell to bash, since the default shell may differ depending on the environment.