Add console endowment (#1355)
* add console endowment file

* add console endowment and test, updated tests to accomodate new endowment

* replaced usage of snapName with snapId

* address comments and fix endowments test

* fix coverage

* actually fix coverage

* Refactor console endowment to include a subset of available console functions (#1361)

* Refactor console endowment to include subset of available console values

The console endowment currently includes the all console functions, which causes SES hardening to fail for some reason. By limiting it to a subset of the available values, we can prevent this hardening issue.

* Remove unavailable console method

The 'createTask' console method is unavailable in Node.js, so it was removed.

* Fix console endowment test

* Update coverage values for lines, branches, functions and statements

The coverage metrics were updated with the new values. The branches and lines coverage have slightly increased, and the functions and statements coverage has a small increase too.

* updated tests

* addressed comments

* fix type errors

* revert changes with EndowmentsWithFactoryOptions

* fix test

* update coverage

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>