BREAKING: Add `create-snap-cli` command (#1268)
Separate the `snaps-cli` in two commands. The `init` command moves to its own package, which can be used directly with `yarn/npm create`. The rest of the commands stay in the `snaps-cli` package. The reasoning for this is that all `mm-snap` sub-command other than `init` are used to _manage_ a snap vs. create one.

Rather than having to install the `mm-snap` CLI globally, which could be error prone, the user just runs one command to create a snap, similar to Create React App.

Fixes #1236