Make `manageAccounts` arguments extend `RestrictedMethodParameters`
The arguments of a restricted method must extend the
`RestrictedMethodParameters` type, which is defined as:

  Json[] | Record<string, Json> | void

However, the current definition of `SnapMessageStruct` is not a valid
JSON value due to the presence of an optional field. This optional field
is incorrectly inferred by superstruct as `params?: ... | undefined`.

This PR adds a workaround to make `SnapMessageStruct` a valid JSON using
the following pattern:

  { A, B? } -> { A } | { A, B }

See: https://github.com/ianstormtaylor/superstruct/issues/983