Delegate call to `saveSnapKeyring` to `SnapKeyring`
The `rpc-methods` repo should not have knowledge about the logic behind
the `SnapKeyring` nor decide when its state needs to be saved.

This PR passes the `saveSnapKeyring` callback function to `SnapKeyring`,
and stops checking the methods being called.

The `manageAccounts` method is still 100% covered, but since some code
was removed, the global coverage threshold decreased.