Check minimum Node version in CLI (#1797)
We got several reports about the CLI not working, which was caused by an
outdated Node.js version. Some functionality we use requires at least
Node.js 16.17.0 or Node.js 18.6.0. Node.js <16 is not support at all.

To prevent this, I've added a check to the CLI, to make sure that the
user knows what the problem is.