Use browser native SHA256 when possible (#1953)
Use browser native SHA256 whenever possible as it is WAY faster in our
environment.

Example benchmark (-96% 😎):
```
noble SHA256 of 143134685 bytes took 10131.300000190735 ms
browser SHA256 of 143134685 bytes took 351.90000009536743 ms
```

Because of this, the checksum function is now async. All functions that
do checksum validation must now also be async. This is the main chunk of
changes in this PR.

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>