Improve `SnapController` constructor typing (#2023)
Improve SnapController constructor typing to improve compatibility with
mobile. Most of the constructor arguments have sane defaults provided by
the SnapController itself and are therefore allowed to be undefined. The
feature flags are also allowed to be true, false and undefined.

These issues weren't caught until now because the extension integration
is not typed.