Refactor `NpmLocation` class (#2038)
Refactor `NpmLocation` to move most logic to a `BaseNpmLocation` class
that can be extended in the mobile client. A class extending the base
class only need to implement `fetchNpmTarball` reducing code
duplication. As part of this refactor I've also simplified the
implementation in multiple places and improved the coverage.

Fixes https://github.com/MetaMask/snaps/issues/2040

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>