Revert usage of `DecompressionStream` (#2052)
Following the merge of #1971 we saw increased failures when fetching
snaps. We haven't been able to fully understand why yet, but it looks
like it may be a race condition in some of the stream code. Out of
caution, we revert part of the PR to stop using `DecompressionStream`
for now.