Move offscreen logic to shared proxy classes (#2093)
Moves most offscreen logic to a shared proxy executor and execution
service that can be consumed by the offscreen executor. This is also
useful for mobile. This PR is very similar to #2083 but targeting main
to simplify the mobile PR.

This PR also removes the notion of a passed `frameUrl` in proxy
executors. Instead they use a constructor argument that defaults to a
hardcoded URL based on their version.

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>