BREAKING: Implement testing framework using Node.js executor (#1982)
This changes the implementation of `snaps-jest` to use a Node.js-based
executor, and remove the use of a (headless) browser to run tests. This
removes a lot of overhead of running a browser, and significantly speeds
up running Snaps.

## Changes

- **BREAKING**: The network mocking functionality was removed.
- This might be reimplemented in the future, if we can find a good way
to do this.
   - Ethereum JSON-RPC mocking is not affected by this change.
- **BREAKING**: `mockJsonRpc` no longer returns a Promise.
- If you're using it with `await` this doesn't break anything (since you
can wait any value), though it's now redundant, and recommended to
remove it.
- You can now pass more options to the `installSnap` function, for
example, to specify the current locale, initial state, or secret
recovery phrase.
- The `close` function on the installed Snap object is now deprecated.
Snaps are automatically "closed" when a new Snap is being installed, or
when all tests are finished.