Always create dynamic UI for returned content (#2152)
This PR makes changes to always create dynamic UI for returned static
content from handlers. It introduces a `transformSnapRpcRequestResult`
function that can be used in the future for any transformation needed
for snap results before returning them.

This removes the need for the `SnapController` to handle phishing
detection and extended validation of the UI. The
`SnapInterfaceController` now manages this for all handler types and
inputs.

This PR also adds an assertion to check that returned interface ids are
valid.