[BREAKING] Move endowments from Controllers to RPC methods (#2155)
This PR is a **BREAKING** change because certain elements are moved from
one package to another and import across the libraries is affected.

### Summary of changes
All endowment related functionalities and specifications are moved from
the `endowments` folder within `snaps-controllers` to the `endowments`
folder located under `snaps-rpc-methods` package and exported from
there.
Export of certain functions from endowments required adjustments to the
index files, so these changes are made as well.
Fixed are related imports of the endowments in another packages affected
(e.g. snaps-simulator).
Code coverage thresholds are adjusted to the new calculations since
there are no functional or logical changes.

### Reasons for the changes
New features being implemented require endowment related specifications
and functionalities, in different packages.

The example of this is Dynamic Permissions feature which require
processing of the raw permission specification in order to request
permission grant to the Permission Controller. This process requires
specific utility functions as well as the caveat mappers or related
specifications which cannot be imported from `snaps-controllers` package
directly into the `snaps-rpc-methods` package.

The best solution for now is to move the endowment related code to the
`snaps-rpc-methods` package and export it from there.