Fix issue installing non-allowlisted Snaps in allowlist mode (#2196)
Fixes an issue installing non-allowlisted Snaps in allowlist mode.
Previously we would fail if resolving a version range to an allowlisted
version would fail. We cannot do this anymore as some Snaps may still be
available for installation in allowlist mode without the version being
on the allowlist. This PR alters the `resolveVersion` function to return
the unaltered version range if it isn't possible to narrow down the
version to install.

Snaps that require allowlisting will still fail in the
`assertIsInstallAllowed` step which follows this step. This issue was
not caught before testing in the extension since the SnapController
tests run with a mocked registry.