Improve timeout handling (#2242)
This PR improves timeout handling in the SnapController and in the
execution services, it makes a handful of changes:
- Adds a timeout of `2` seconds to the `ping` request, sent to the
execution environment after boot. This lets us fail more immediately if
the execution environment has loaded but is non-responsive (e.g. if the
execution env is down, the iframe loads a Chrome failure page, which
still counts as the iframe loading).
- Adds `maxInitTime` as a constructor argument allowing for better
configuration of how long Snaps (and the execution environment) has in
total to boot before timing out. Previously `maxRequestTime` was used
for the initialization timeout.
- Changes all of the error messaging around timeouts in an attempt to
give more information than the previous vague error message: "The
request timed out.".