Handle unavailable registry more gracefully (#2256)
This PR makes a few changes to the registry controller implementation to
attempt to handle an unavailable registry (e.g. invalid signature) more
gracefully.

Instead of throwing when the registry fails to update or populate, we
return `SnapsRegistryStatus.Unavailable`. This status is meant to
indicate to the `SnapController` that we can't currently access the
latest registry and thus our data might be out of date. For Snaps that
don't require allowlisting, this shouldn't be a problem and they will
still be installable. If a user is trying to install a Snap that does
require allowlisting they are now greeted with a slightly better error
message.

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>