Allow `maxRequestTime` on `endowment:rpc` (#2291)
When implementing `maxRequestTime`, validation support for the property
on `endowment:rpc` in the manifest was left out accidentally. This would
cause problems when devs try to use `maxRequestTime` with the endowment
permission. This PR fixes that by making the endowment use similar
validation to the other handler endowments.

Fixes https://github.com/MetaMask/snaps/issues/2290