BREAKING: Refactor encryption to enable caching (#2316)
This PR is a refactor of the Snap state encryption logic, it moves the
logic from the manageState RPC method implementation to the
SnapController. It re-implements the encryption/decryption logic using a
different set of methods, as the previously used implementation had
encapsulated the encryption key. With this refactor it is now possible
to cache the encryption key used by the Snap in the runtime, making
continuous requests to update the state much faster (numbers TBD).

With this refactor it is now the responsibility of the SnapController to
manage encrypting/decrypting the Snap state and thus the method hooks
`getSnapState` and `updateSnapState` have had breaking changes, they are
now asynchronous and expected to fully handle encryption/decryption.
Because of that, this change required changes to both snaps-jest and
snaps-simulator for compatibility.

This will not work as-is on mobile, since the encryptor on mobile does
not have feature parity with `browser-passworder`. This will need to be
addressed before we can use this change on mobile.

Closes #2315

---------

Co-authored-by: Maarten Zuidhoorn <maarten@zuidhoorn.com>