Fix encryption key caching issues (#2326)
This PR fixes a few problems discovered when trying to integrate the new
encryption caching mechanism.

When the Snap is initially installed, its state will be `undefined`.
Thus, `getSnapState` needs to account for this when deciding when to
return `null`, which is the normalized value we use for lack of state.

Also in our tests, the salt is always the same, so tests were passing,
but in reality the salt changes every time `generateSalt` is called. For
that reason, we cache the salt alongside the encryption key for a
session.