fix: properly tear down partially initialized executors (#2348)
Refactor execution timeout logic to fix a problem where the execution
worker would not be torn down correctly when failing to initialize. This
PR also moves the entire initialization timeout logic to the
`AbstractExecutionService`, simplifying the `SnapController`. Moving the
timeout logic fixes problems where job initialization was not cancelled
properly resulting in "Snap is already being executed", since execution
of the async functions inside `AbstractExecutionService` would continue
even after the timeout.

The ExecutionService gets an optional `maxInitTime` constructor argument
that can be set to define init time for the specific service. This
timeout value is spread between initiating the execution job and running
the first execution of the Snap. If the execution environment fails to
initiate we attempt to tear it down in case it has partially started
(this is at least the case for iframes). We make sure that the timeout
is applied to `initEnvStream` which has the smallest possible footprint
and worry for side-effects. Additionally in case an error happens when
terminating a Snap we catch that error an move on.