fix: unintentionally low timeout value for proxy executors (#2389)
Fixes a problem with proxy execution environments having unintentionally
low timeout values. The assumption is that `initJob` resolves once the
execution environment is perceived to be ready for requests and
therefore we ping the execution env with a very low timeout to detect
issues and fail early. However, `initJob` was resolving way before the
actual proxy executor was ready. This problem can be fixed by
effectively pinging the execution environment as part of the
initialization process of proxy executors as this loads the proxy
executor and doesn't resolve until the executor responds.

Fixes #2372